requirements.txt
python-dotenv==1.0.0
pathlib==1.0.1
typing==3.7.4.3
pytest==7.4.0
black==23.3.0
flake8==6.0.0